Move and rename handler files
